+++ Tests/Unit/Domain/Model/Event/SingleEventTest.php
@@ -145,7 +145,7 @@ final class SingleEventTest extends UnitTestCase
      */
     public function setStartSetsStart(): void
     {
-        $date = new \DateTimeImmutable();
+        $date = new \DateTime();
         $this->subject->setStart($date);
 
         self::assertSame($date, $this->subject->getStart());
@@ -174,7 +174,7 @@ final class SingleEventTest extends UnitTestCase
      */
     public function setEndSetsEnd(): void
     {
-        $date = new \DateTimeImmutable();
+        $date = new \DateTime();
         $this->subject->setEnd($date);
 
         self::assertSame($date, $this->subject->getEnd());
@@ -203,7 +203,7 @@ final class SingleEventTest extends UnitTestCase
      */
     public function setEarlyBirdDeadlineSetsEarlyBirdDeadline(): void
     {
-        $date = new \DateTimeImmutable();
+        $date = new \DateTime();
         $this->subject->setEarlyBirdDeadline($date);
 
         self::assertSame($date, $this->subject->getEarlyBirdDeadline());
@@ -232,7 +232,7 @@ final class SingleEventTest extends UnitTestCase
      */
     public function setRegistrationDeadlineSetsRegistrationDeadline(): void
     {
-        $date = new \DateTimeImmutable();
+        $date = new \DateTime();
         $this->subject->setRegistrationDeadline($date);
 
         self::assertSame($date, $this->subject->getRegistrationDeadline());
@@ -567,7 +567,7 @@ final class SingleEventTest extends UnitTestCase
      */
     public function setRegistrationStartSetsRegistrationStart(): void
     {
-        $date = new \DateTimeImmutable();
+        $date = new \DateTime();
         $this->subject->setRegistrationStart($date);
 
         self::assertSame($date, $this->subject->getRegistrationStart());